Support window.open links
